Fix byte literal cast in shouldRejectWrongVersionByte test
0x99 exceeds signed byte range, needs explicit (byte) cast.
Found during standalone test execution (58/58 tests passing).

+++ app/src/test/java/com/atakmap/android/meshtastic/encryption/AppLayerEncryptionManagerTest.java
@@ -382,7 +382,7 @@ class AppLayerEncryptionManagerTest {
         byte[] encrypted = manager.encrypt("test".getBytes(StandardCharsets.UTF_8));
 
         // When - change version byte to unsupported version
-        encrypted[1] = 0x99;
+        encrypted[1] = (byte) 0x99;
 
         // Then
         byte[] result = manager.decrypt(encrypted);
